refactor: move collector contents logic to CollectorContents from CollectorMainPage (#1326)
* refactor: rename to CollectorMainPage


* refactor: move collector contents logic to CollectorContents from CollectorMainPage


* chore: remove console.debug


---------
